Allow renew-cron to remove cronjob if it was added earlier
If the `certbot_auto_renew` variable was `True`, and the cronjob was
added, setting it to `False` would not remove it. By always including
the `renew-cron.yml` task file, and instead checking the
`certbot_auto_renew` value there, and pick a value for `state` based on
it, we can clean up a cronjob if it is no longer wanted.