WIP: Adjust defaults for webroot arg and stopping services
We need to have a default (empty) value for `certbot_webroot_arg` to
avoid ansible errors.

Have an empty list of services to stop for standalone operation, and
leave it to the user to decide entirely. I believe it is preferred for
certificates to fail renewal than accidentally stopping a service that
is not supposed to be stopped.